Added annotation mode with json reload.
Still need to make censors selectable, draggable and resizable
