+++ renderer/player.html
@@ -2,13 +2,13 @@
 <html>
 <head>
   <link rel="stylesheet" type="text/css" href="style.css">
-  <link rel="shortcut icon" href="filmstrip.png">
+  <link rel="shortcut icon" href="../resources/filmstrip.png">
   <title>BYU International Cinema - Video Player</title>
 </head>
 <body>
   <div id='splashScreen'>
     <div class='container'>
-      <image id='logo' src='ic_round.png'></img><br>
+      <img id='logo' src='../resources/ic_round.png'></img><br>
       <a href='#' id='playButton' onclick='player.button_press()'>
         <!-- Generator: Adobe Illustrator 19.0.0, SVG Export Plug-In  -->
         <svg version="1.1"
@@ -29,18 +29,18 @@
       Select Video Files: <input type="file" id="filePicker" multiple>
     </div>
   </div>
-  
+
   <div id='playerContainer'>
     <button id='returnBtn' onclick='player.hide_player()'>&#10094;&nbsp;Return to Menu</button>
+    <button id='reloadJsonBtn' onclick='player.reload_json()'>Reload JSON Filters</button>
     <video id='player' width='1500' height='800' controls="true">
     </video>
 
-    
     <div id='box'></div>
   </div>
 
   <script type="text/javascript" src="eventlisteners.js"></script>
-  <script type="text/javascript" src="script.js"></script>
+  <script type="text/javascript" src="renderer.js"></script>
 
 </body>
 </html>
