+++ renderer/style.css
@@ -177,7 +177,7 @@ h1{
   position: absolute;
   top: 0;
   left: 2px;
-  z-index: 2;
+  z-index: 3;
   background-color: rgba(255,255,255, 0.2);
   color: rgba(255,255,255,0.5);
   border: 1px solid black;
@@ -199,6 +199,31 @@ h1{
   border: 1px solid rgba(255,255,255, 0.2);
 }
 
+#reloadJsonBtn {
+  position: absolute;
+  top: 0;
+  right: 2px;
+  z-index: 3;
+  background-color: rgba(255,255,255, 0.2);
+  color: rgba(255,255,255,0.5);
+  border: 1px solid black;
+  outline: none;
+  padding: 15px;
+  font-size: 10px 15px;
+  transition: all .25s;
+}
+
+#reloadJsonBtn:hover {
+  color: rgba(255,255,255,1);
+  border: 1px solid grey;
+}
+
+#reloadJsonBtn:active {
+  color: rgba(255,255,255,0.5);
+  background-color: rgba(255,255,255, 0.2);
+  border: 1px solid rgba(255,255,255, 0.2);
+}
+
 video {
   width: 100%;
   height: 100%;
@@ -212,13 +237,13 @@ video::-webkit-media-text-track-display {
 
 /*Styling the video controls panel*/
 video::-webkit-media-controls-panel {
-  z-index: 100;
+  /*z-index: 100;*/
 }
 
 
 #box {
   /* Set z-index to MAX VALUE */
-  z-index: 2147483647;
+  z-index: 2/*2147483647*/;
   border: 0px solid red;
   position: absolute;
   top: 0;
@@ -228,6 +253,20 @@ video::-webkit-media-controls-panel {
 }
 
 .censor {
-  z-index: 2147483647;
+  z-index: 3/*2147483647*/;
   position: relative;
+}
+
+.censor-annotate:hover::after {
+  position: absolute;
+  display: block;
+  content: '';
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
+  background: #4c88ff;
+  opacity: 0.50;
+  z-index: 4;
+  resize: both;
 }
