+++ resources/example/KungFuPandamv_IC.json
