+++ resources/kung-test/KungFuPandamv_IC.icf
